Update grex.sh to run task set properly over ssh
Taskset doesn't use the same file for accessing `PATH` when run over ssh rather than directly on the machine. This causes it to not see `poetry` when starting T2, causing the startup to fail. I added a line that explicitly adds the parent directory for `poetry` (export PATH="$HOME/.local/bin:$PATH") in the t2_cmd function, which fix the issue with starting up grex over ssh.