Update grex.sh to export `~/.local/bin` in t2_cmd
Change from previous PR because the `export PATH` bit was exterior to `echo`, making it useless. The goal of the PR is to make sure that `~/.local/bin` is available to any subshells the command string runs in so that `poetry` is on the `PATH`. This should address a issue I've had where running grex over ssh terminates when trying to run poetry for T2.